+++ frontend/package.json
@@ -5,19 +5,26 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "buildScss": "node-sass -r -o app app",
-    "buildTs": "tsc -p .",
-    "buildTs:w": "npm run buildTs -- -w",
-    "buildScss:w": "node-sass -r -w -o ./app ./app",
-    "build:w": "npm run buildScss:w & npm run buildTs:w",
-    "bs": "browser-sync start --config bs-config.js",
-    "serve": "npm run bs & npm run build:w"
+    "build": "webpack --config webpack.prod.js --mode 'production'",
+    "serve": "webpack-dev-server --config webpack.dev.js"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
     "@types/jquery": "^3.3.1",
     "@types/jquery-mousewheel": "^3.1.8",
-    "node-sass": "^4.8.3"
+    "clean-webpack-plugin": "^0.1.19",
+    "css-loader": "^0.28.11",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "file-loader": "^1.1.11",
+    "html-webpack-plugin": "^3.2.0",
+    "node-sass": "^4.8.3",
+    "sass-loader": "^6.0.7",
+    "style-loader": "^0.20.3",
+    "ts-loader": "^4.2.0",
+    "webpack": "^4.5.0",
+    "webpack-cli": "^2.0.14",
+    "webpack-dev-server": "^3.1.3",
+    "webpack-merge": "^4.1.2"
   }
 }
