+++ frontend/src/index.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>futter@siemens</title>
+
+  <link rel="shortcut icon" href="favicon.ico" type="image/x-icon" />
+
+  <script src="https://cdn.polyfill.io/v2/polyfill.min.js"></script>
+
+  <script src="https://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4="
+    crossorigin="anonymous"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-mousewheel/3.1.13/jquery.mousewheel.min.js"></script>
+
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/flag-icon-css/2.8.0/css/flag-icon.min.css" />
+
+  <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN"
+    crossorigin="anonymous">
+</head>
+
+<body>
+  <div id='site'>
+    <header></header>
+    <div id='plan' class="daysContainer"></div>
+    <footer>
+      <div class="content"></div>
+      <div class="spacer"></div>
+      <div>
+        <a href="https://paypal.me/muxe/1">
+          <div class="link">
+            <div>
+              <i class="fa fa-paypal" aria-hidden="true"></i>
+            </div>
+            <div>donate</div>
+          </div>
+        </a>
+      </div>
+    </footer>
+  </div>
+</body>
+
+</html>
