+++ frontend/src/style.scss
@@ -1,0 +1,209 @@
+body,
+html {
+  margin: 0;
+  padding: 0;
+  height: 100%;
+  overflow: auto;
+}
+
+body {
+  font-family: 'Siemens sans', 'Open Sans', Arial, sans-serif;
+  font-size: 1em;
+  line-height: 1.5em;
+  background: #555f69;
+  // color: #3c464b;
+}
+
+::-webkit-scrollbar {
+  width: 12px;
+  height: 12px;
+}
+
+::-webkit-scrollbar-track {
+  background: #788791;
+}
+
+::-webkit-scrollbar-thumb {
+  background: rgb(15, 130, 135);
+}
+
+// ::-webkit-scrollbar-button {
+//     background: green;
+// }
+
+#site {
+  height: 100%;
+  display: grid;
+  grid-template-rows: auto 1fr auto;
+  grid-template-columns: 1fr auto 1fr;
+
+  & > header {
+    grid-row: 1;
+    grid-column: 1/4;
+    padding: 10px;
+    background: #2d373c;
+    color: white;
+    display: flex;
+
+    .flag-icon {
+      margin-left: 5px;
+      cursor: pointer;
+    }
+
+    span:nth-child(1) {
+      flex-grow: 1;
+    }
+  }
+
+  & > #plan {
+    grid-row: 2;
+    grid-column: 1/4;
+  }
+
+  & > footer {
+    grid-row: 3;
+    grid-column: 1/4;
+    padding: 8px;
+    background: #2d373c;
+    color: white;
+
+    align-items: center;
+
+    display: flex;
+    .spacer {
+      flex-grow: 1;
+    }
+
+    a {
+      div.link {
+        display: flex;
+        align-items: center;
+        background: rgba(255, 255, 255, 0.1);
+        border-radius: 10px;
+        padding: 3px 10px;
+
+        &:hover {
+          background: rgba(255, 255, 255, 0.2);
+        }
+      }
+      i {
+        font-size: 1em;
+        margin-right: 10px;
+      }
+      color: white;
+    }
+  }
+}
+
+.daysContainer {
+  display: flex;
+  overflow-x: auto;
+  overflow-y: auto;
+  padding: 10px;
+}
+
+// $color1: #f5f5f0;
+// $color2: #ebebe3;
+// $textColor: #2d373c;
+// $header: #41aaaa;
+
+$color1: rgb(15, 130, 135);
+$color2: rgb(35, 145, 150);
+$textColor: white;
+$header: rgb(60, 70, 75);
+
+.weekend {
+  min-width: 10px;
+  min-height: 10px;
+  background: #3c464b;
+}
+
+.day {
+  display: grid;
+  grid-template-rows: auto 1fr 1fr 1fr 1fr 1fr;
+  flex-grow: 1;
+
+  & + &,
+  & + .weekend,
+  .weekend + & {
+    margin-left: 10px;
+  }
+
+  header {
+    min-width: 20em;
+    background: $header;
+    color: white;
+    padding: 10px;
+  }
+  .meal {
+    color: $textColor;
+    padding: 10px;
+    display: flex;
+    flex-direction: column;
+    &:nth-child(even) {
+      background: $color1;
+    }
+    &:nth-child(odd) {
+      background: $color2;
+    }
+
+    .spacer {
+      flex-grow: 1;
+    }
+
+    .description {
+      flex-grow: 1;
+    }
+
+    .prices {
+      margin-top: -1.5em;
+      display: flex;
+      flex-direction: row;
+      align-items: right;
+    }
+
+    .price {
+      text-align: right;
+      font-style: italic;
+      &::after {
+        content: ' €';
+      }
+    }
+
+    .price + .price {
+      margin-left: 10px;
+      font-size: 0.9em;
+      line-height: 1.9em;
+      vertical-align: bottom;
+    }
+  }
+}
+
+@media screen and (max-width: 500px) {
+  body {
+    font-size: 1em;
+  }
+
+  .daysContainer {
+    flex-direction: column;
+    padding: 0;
+  }
+
+  .day {
+    display: block;
+  }
+
+  .day + .day,
+  .weekend + .day,
+  .day + .weekend {
+    margin: 0;
+  }
+
+  .day + .weekend {
+    margin-top: 10px;
+  }
+
+  .weekend + .day {
+    margin-top: 10px;
+  }
+}
