+++ frontend/webpack.dev.js
@@ -1,0 +1,12 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  devtool: 'inline-source-map',
+  devServer: {
+    contentBase: './dist',
+    port: 9000,
+    historyApiFallback: true
+  },
+  mode: 'development'
+});
