+++ frontend/webpack.prod.js
@@ -1,0 +1,4 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {});
